AI-14-cmd/Hackathon-GT/requirements.txt
pandas
matplotlib
fpdf2
openai
google-generativeai
python-dotenv
